Split out utils.rs
